Week8/Day2/DailyChallenge/phone_number/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.1.2
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
joblib==1.2.0
phonenumbers==8.13.0
Pillow==9.2.0
tqdm==4.64.1
tzdata==2022.5